+++ include/math_utils.h
@@ -2,12 +2,13 @@
 
 #include "vec3.h"
 #include "quaternion.h"
+#include "float_type.h"
 
 struct Transform;
 
 namespace MathUtils
 {
-    double degrees_to_radians(double degrees);
+    FloatType degrees_to_radians(FloatType degrees);
 
     Vec3 rotate(const Quaternion &q, const Vec3 &v);
