+++ include/ray.h
@@ -9,7 +9,7 @@ struct Ray
 
     Ray(const Point3 &origin, const Vec3 &direction) : origin(origin), direction(direction) {}
 
-    Point3 at(double t) const
+    Point3 at(FloatType t) const
     {
         return origin + t * direction;
     }
