+++ src/math_utils.cpp
@@ -6,16 +6,16 @@
 namespace MathUtils
 {
 
-    double degrees_to_radians(double degrees)
+    FloatType degrees_to_radians(FloatType degrees)
     {
-        return degrees * std::numbers::pi / 180.0;
+        return degrees * static_cast<FloatType>(std::numbers::pi) / static_cast<FloatType>(180.0);
     }
 
     Vec3 rotate(const Quaternion &q, const Vec3 &v)
     {
         Vec3 u(q.x, q.y, q.z);
-        double s = q.w;
-        return 2.0 * Vec3::dot(u, v) * u + (s * s - Vec3::dot(u, u)) * v + 2.0 * s * Vec3::cross(u, v);
+        FloatType s = q.w;
+        return static_cast<FloatType>(2.0) * Vec3::dot(u, v) * u + (s * s - Vec3::dot(u, u)) * v + static_cast<FloatType>(2.0) * s * Vec3::cross(u, v);
     }
 
     Vec3 transform_point(const Transform &t, const Vec3 &point)
@@ -30,8 +30,8 @@ namespace MathUtils
         Vec3 right = Vec3::normalize(Vec3::cross(world_up, forward));
         Vec3 up = Vec3::cross(forward, right);
 
-        double yaw = std::atan2(forward.x, forward.z);
-        double pitch = std::asin(-forward.y);
+        FloatType yaw = std::atan2(forward.x, forward.z);
+        FloatType pitch = std::asin(-forward.y);
         Quaternion rotation = Quaternion::from_euler(0, pitch, yaw);
 
         return Transform(position, rotation);
