+++ src/vec3.cpp
@@ -9,15 +9,15 @@ Vec3 operator-(const Vec3 &u, const Vec3 &v)
 {
     return Vec3(u.x - v.x, u.y - v.y, u.z - v.z);
 }
-Vec3 operator*(double t, const Vec3 &v)
+Vec3 operator*(FloatType t, const Vec3 &v)
 {
     return Vec3(t * v.x, t * v.y, t * v.z);
 }
-Vec3 operator*(const Vec3 &v, double t)
+Vec3 operator*(const Vec3 &v, FloatType t)
 {
     return Vec3(t * v.x, t * v.y, t * v.z);
 }
-Vec3 operator/(const Vec3 &v, double t)
+Vec3 operator/(const Vec3 &v, FloatType t)
 {
     return Vec3(v.x / t, v.y / t, v.z / t);
 }
